requirements.txt
async-timeout==4.0.2
beautifulsoup4==4.11.1
chardet==3.0.4
Deprecated==1.2.13
easygui==0.96
grpcio==1.46.3
importlib-metadata==4.11.4
meld==3.20.0
numpy==1.16.2
packaging==21.3
pandas==0.23.3+dfsg
pika==1.2.1
protobuf==4.21.1
pudb==2018.1
pycairo==1.16.2
Pygments==2.3.1
PyGObject==3.30.4
pyparsing==3.0.9
python-apt==1.8.4.3
python-dateutil==2.7.3
python-debian==0.1.35
pytz==2019.1
redis==4.3.2
six==1.12.0
soupsieve==2.3.2.post1
typing_extensions==4.2.0
urwid==2.0.1
wrapt==1.14.1
zipp==3.8.0
